feat: Use Request from new Snuba SDK
The new version of the Snuba SDK sends a Request object to Snuba instead of the
Query class. This allows separating out flags that are not specific to the Query
and makes the Query interface easier to use. It also adds a mandatory app_id to
the Request, but that can be left as "default" until it starts getting used.

The actual format of the JSON sent to Snuba has not changed, so this change
should have no impact on the actual queries being run. It is simply moving some
fields into a new class.